% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREDICTv3.R
\name{PREDICTv3}
\alias{PREDICTv3}
\title{Breast Cancer Prognosis Prediction Tool}
\usage{
PREDICTv3(
  year,
  age,
  smoker,
  er,
  pr,
  her2,
  ki67,
  size,
  grade,
  screen,
  nodes,
  radio,
  heart.gy,
  horm,
  gen,
  traz,
  bis,
  censor
)
}
\arguments{
\item{year}{The year of the patient's diagnosis.}

\item{age}{The patient's age at the time of diagnosis.}

\item{smoker}{Indicates whether the patient has a history of smoking.}

\item{er}{The patient's estrogen receptor status (positive or negative).}

\item{pr}{The patient's progesterone receptor status (positive or negative).}

\item{her2}{The status of the patient's human epidermal growth factor receptor 2 (HER2).}

\item{ki67}{The status of the patient's Ki-67 protein, a marker of cell proliferation.}

\item{size}{The size of the patient's tumor, usually measured in centimeters.}

\item{grade}{The grade of the tumor, indicating its aggressiveness.}

\item{screen}{The method used for disease detection (e.g., screening or symptomatic).}

\item{nodes}{The status of lymph nodes (e.g., number of positive nodes).}

\item{radio}{Indicates whether the patient has received radiotherapy.}

\item{heart.gy}{The radiation dose (measured in Gray) delivered to the heart during radiotherapy.}

\item{horm}{Indicates whether the patient has received hormone therapy.}

\item{gen}{The generation of radiotherapy received by the patient (e.g., first, second).}

\item{traz}{Indicates whether the patient has received trastuzumab.}

\item{bis}{Indicates whether the patient has received bisphosphonates.}

\item{censor}{The number of years to consider since surgery (e.g., 5, 10, or 15 years).}
}
\description{
Predicts the survival outcome for a patient based on their clinical information.
}
\examples{
data(toyData)
year <- toyData[1, 'year']
age <- toyData[1, 'age']
smoker <- toyData[1, 'smoker']
er <- toyData[1,'er']
pr <- toyData[1, 'pr']
her2 <- toyData[1, 'her2']
ki67 <- toyData[1, 'ki67']
size <- toyData[1, 'size']
grade <- toyData[1, 'grade']
screen <- toyData[1, 'screen']
nodes <- toyData[1, 'nodes']
radio <- toyData[1, 'radio']
heart.gy <- toyData[1, 'heart_gy']
horm <- toyData[1, 'horm']
gen <- toyData[1, 'gen']
traz <- toyData[1, 'traz']
bis <- toyData[1, 'bis']
censor <- 10
PREDICTv3(year=year,  age=age, screen=screen, size=size, grade=grade, nodes=nodes, er=er, her2=her2, ki67=ki67, pr=pr, radio=radio, heart.gy=heart.gy, horm=horm, gen=gen, traz=traz, bis=bis, smoker = smoker, censor=censor)
}
