% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toyData.R
\docType{data}
\name{toyData}
\alias{toyData}
\title{toyData for}
\format{
A R data frame with 50 rows and 18 variables
\describe{
  \item{clinical data}{Numeric clinical matrix with each
  rows as a different individual and each column as a clinical feature.}}
}
\source{
<https://www.github.com/pengpclab/PREDICTv3>
}
\usage{
data(toyData)
}
\description{
A simulated data set with clinical information for 50 breast cancer cases
}
\examples{
data(toyData)
}
\keyword{datasets}
